Made the plugin load and panel visible
Modified to allow the loading of the plugin.

Also enabled the display of the Ack panel, it shows results.

Missing: selecting the correct line in an OakDocument. Needs more refactoring.

Reduced amount of compiler warnings.

+++ source/controllers/AckMatePlugin.h
@@ -4,7 +4,7 @@
 #import <Cocoa/Cocoa.h>
 
 @protocol TMPlugInController
-- (float)version;
+- (CGFloat)version;
 @end
 
 @interface AckMatePlugin : NSObject
